Start moving command-line processing to libsupport.a
This only moves things, no functional change yet.
