+++ src/support/command-line.h
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2015 WebAssembly Community Group participants
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+//
+// Command line helpers.
+//
+
+#ifndef wasm_command_line_h
+#define wasm_command_line_h
+
+#include "wasm.h"
+
+namespace wasm {
+
+struct Options {
+  // standard options
+  bool debug;
+  std::string infile;
+  std::string outfile;
+  // extra options
+  std::map<std::string, const char *> extra;
+  Options() : debug(false) {}
+};
+
+void processCommandLine(int argc, const char *argv[], Options *options,
+                        const char *help);
+
+}  // namespace wasm
+
+#endif // wasm_command_line_h
